+++ locale/programs/linereader.c
@@ -381,7 +381,7 @@ get_toplvl_escape (struct linereader *lr)
       bytes[nbytes++] = byte;
     }
   while (ch == lr->escape_char
-	 && nbytes < sizeof (lr->token.val.charcode.bytes));
+	 && nbytes < (int) sizeof (lr->token.val.charcode.bytes));
 
   if (!isspace (ch))
     lr_error (lr, _("garbage at end of character code specification"));
